+++ addon/prefs.js
@@ -1,4 +1,3 @@
 pref("extensions.zotero-citegraph.showGhosts", false);
 pref("extensions.zotero-citegraph.sizeBy", "cited-here");
 pref("extensions.zotero-citegraph.colorBy", "year");
-pref("extensions.zotero-citegraph.selftest", false);